+++ services/web/server/src/simcore_service_webserver/login/_2fa_api.py
@@ -20,7 +20,7 @@ from twilio.base.exceptions import TwilioException  # type: ignore[import-untype
 from twilio.rest import Client  # type: ignore[import-untyped]
 
 from ..login.errors import SendingVerificationEmailError, SendingVerificationSmsError
-from ..products.products_models import Product
+from ..products.models import Product
 from ..redis import get_redis_validation_code_client
 from .utils_email import get_template_path, send_email_from_template
