+++ services/web/server/src/simcore_service_webserver/login/_registration_handlers.py
@@ -19,7 +19,7 @@ from servicelib.utils import fire_and_forget_task
 from .._meta import API_VTAG
 from ..constants import RQ_PRODUCT_KEY
 from ..products import products_web
-from ..products.products_models import Product
+from ..products.models import Product
 from ..security.api import check_password, forget_identity
 from ..security.decorators import permission_required
 from ..session.api import get_session
