+++ services/web/server/src/simcore_service_webserver/products/_rest.py
@@ -21,7 +21,7 @@ from ..login.decorators import login_required
 from ..security.decorators import permission_required
 from ..utils_aiohttp import envelope_json_response
 from . import _service, products_web
-from .products_models import Product
+from .models import Product
 
 routes = web.RouteTableDef()
