+++ services/web/server/src/simcore_service_webserver/products/_web_events.py
@@ -16,7 +16,7 @@ from simcore_postgres_database.utils_products import (
 from ..constants import APP_PRODUCTS_KEY, FRONTEND_APP_DEFAULT, FRONTEND_APPS_AVAILABLE
 from ..db.plugin import get_database_engine
 from ._repository import get_product_payment_fields, iter_products
-from .products_models import Product
+from .models import Product
 
 _logger = logging.getLogger(__name__)
