+++ services/web/server/src/simcore_service_webserver/products/_web_middlewares.py
@@ -8,7 +8,7 @@ from servicelib.rest_constants import X_PRODUCT_NAME_HEADER
 
 from .._meta import API_VTAG
 from ..constants import APP_PRODUCTS_KEY, RQ_PRODUCT_KEY
-from .products_models import Product
+from .models import Product
 
 _logger = logging.getLogger(__name__)
