+++ services/web/server/src/simcore_service_webserver/products/products_web.py
@@ -6,9 +6,9 @@ from ._web_helpers import (
 )
 
 __all__: tuple[str, ...] = (
-    "get_product_name",
     "get_current_product",
     "get_current_product_credit_price_info",
+    "get_product_name",
     "get_product_template_path",
 )
 # nopycln: file
