+++ services/web/server/src/simcore_service_webserver/statics/_events.py
@@ -17,7 +17,7 @@ from yarl import URL
 
 from ..application_settings import ApplicationSettings, get_application_settings
 from ..constants import APP_PRODUCTS_KEY
-from ..products.products_models import Product
+from ..products.models import Product
 from ._constants import (
     APP_FRONTEND_CACHED_INDEXES_KEY,
     APP_FRONTEND_CACHED_STATICS_JSON_KEY,
