+++ services/web/server/src/simcore_service_webserver/users/_users_rest.py
@@ -27,7 +27,7 @@ from ..groups import api as groups_api
 from ..groups.exceptions import GroupNotFoundError
 from ..login.decorators import login_required
 from ..products import products_web
-from ..products.products_models import Product
+from ..products.models import Product
 from ..security.decorators import permission_required
 from ..utils_aiohttp import envelope_json_response
 from . import _users_service
