+++ services/web/server/tests/unit/isolated/test_products_middlewares.py
@@ -17,7 +17,7 @@ from simcore_service_webserver.products._web_events import _set_app_state
 from simcore_service_webserver.products._web_middlewares import (
     discover_product_middleware,
 )
-from simcore_service_webserver.products.products_models import Product
+from simcore_service_webserver.products.models import Product
 from simcore_service_webserver.statics._constants import FRONTEND_APP_DEFAULT
 from yarl import URL
