+++ services/web/server/tests/unit/with_dbs/03/invitations/test_invitations.py
@@ -20,7 +20,7 @@ from simcore_service_webserver.invitations.errors import (
     InvalidInvitationError,
     InvitationsServiceUnavailableError,
 )
-from simcore_service_webserver.products.products_models import Product
+from simcore_service_webserver.products.models import Product
 from yarl import URL
